Updating Composer install instructions
The recommended way to add new dependencies to a project via composer is using the `require` command, instead of manually editing the json file. This command runs a more controlled update that only affects the needed packages and is less error prone.

It will also bind to the latest version of the package using the `^` operator which is also recommended.